-Lots of restructuring -Look at routes/api.php for api urls -Worked on creating resources and extending "base" resources to include a default with () function -Edited databases a little more, might need to re-migrate and --seed -Probably errors in some of the code as halfway through I started restructing. -Pairing currently works with an even number of players, need to work out a bye system. -Should be able to register people to the tournaments now.

+++ app/Http/Controllers/Api/TournamentController.php
@@ -4,11 +4,10 @@ namespace App\Http\Controllers\Api;
 
 use Validator;
 use Illuminate\Http\Request;
-use App\Http\Controllers\Controller;
 
 use App\Models\Tournament;
 use App\Http\Resources\Tournament as TournamentResource;
-use App\Http\Resources\Tournaments;
+use App\Http\Resources\Collections\Tournaments;
 
 class TournamentController extends Controller
 {
@@ -21,58 +20,43 @@ class TournamentController extends Controller
     {
         //Validate get request
         $v = Validator::make($r->all(), [
-            'limit' => 'integer',
+            'perPage' => 'required_with:page|integer',
+            'page' => 'sometimes|integer',
         ]);
-        if ($v->fails()) {
-            $errs = $v->errors();
-            $response = [
-                'success' => false,
-                'errors' => $errs->all(),
-                'data' => [],
-            ];
-            return response()
-                ->json($response)
-                ->setStatusCode(400);
-        }
-        $limit = isset($r->limit)?$r->limit:10;
+        if ($v->fails())
+            return $this->errors($v);
 
-        //Get the 5 most recent tournaments
-        return new Tournaments (Tournament::orderBy('event_time', 'desc')->paginate($limit));
+        //Get the 10 most recent tournaments
+        return (new Tournaments (Tournament::orderBy('event_time', 'desc')
+            ->paginate(isset($r->perPage)?$r->perPage:10)))
+            ->response()
+            ->setStatusCode(200);
     }
 
     /**
-     * Store a newly created resource in storage.
+     * Create a new resource in storage.
      *
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $r)
+    public function create(Request $r)
     {
         //Validate post request
         $v = Validator::make($r->all(), [
             'title' => 'required|string',
-            'event_time' => 'date',
-            'game_id' => 'required|exists:games,id',
-            'format_id' => 'required|exists:formats,id',
+            'eventTime' => 'date',
+            'gameId' => 'required|exists:games,id',
+            'formatId' => 'required|exists:formats,id',
         ]);
 
-        if ($v->fails()) {
-            $errs = $v->errors();
-            $response = [
-                'success' => false,
-                'errors' => $errs->all(),
-                'data' => [],
-            ];
-            return response()
-                ->json($response)
-                ->setStatusCode(400);
-        }
+        if ($v->fails()) 
+            return $this->errors($v);
 
         //Create a new tournament and return it's information
         $tournament = new Tournament ();
         $tournament->title = $r->title;
-        $tournament->event_time = isset($r->event_time)?$r->event_time:now();
-        $tournament->game_id = $r->game_id;
-        $tournament->format_id = $r->format_id;
+        $tournament->event_time = isset($r->eventTime)?$r->eventTime:now();
+        $tournament->game_id = $r->gameId;
+        $tournament->format_id = $r->formatId;
         $tournament->save();
 
         return (new TournamentResource($tournament))
@@ -86,48 +70,39 @@ class TournamentController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $r, $id)
+    public function update(Request $r)
     {
         //Validate put request
-        $v = Validator::make(array_merge(['id' => $id], $r->all()), [
+        $v = Validator::make($r->all(), [
             'id' => 'required|exists:tournaments,id',
-            'title' => 'string',
-            'event_time' => 'date',
-            'fk_game_id' => 'exists:games,id',
-            'fk_format_id' => 'exists:formats,id',
+            'title' => 'sometimes|string',
+            'eventTime' => 'sometimes|date',
+            'gameId' => 'exists:games,id',
+            'formatId' => 'exists:formats,id',
         ]);
 
-        if ($v->fails()) {
-            $errs = $v->errors();
-            $response = [
-                'success' => false,
-                'errors' => $errs->all(),
-                'data' => [],
-            ];
-            return response()
-                ->json($response)
-                ->setStatusCode(400);
-        }
+        if ($v->fails()) 
+            return $this->errors($v);
 
-        //Create a new tournament and return it's information
+        //Update an existing tournament and return it's information
         $tournament = Tournament::findOrFail($r->id);
         if (isset($r->title))
             $tournament->title = $r->title;
 
-        if (isset($r->event_time))
-            $tournament->event_time = $r->event_time;
+        if (isset($r->eventTime))
+            $tournament->event_time = $r->eventTime;
 
-        if (isset($r->fk_game_id))
-            $tournament->fk_game_id = $r->fk_game_id;
+        if (isset($r->gameId))
+            $tournament->game_id = $r->gameId;
 
-        if (isset($r->fk_format_id))
-            $tournament->fk_format_id = $r->fk_format_id;
+        if (isset($r->formatId))
+            $tournament->format_id = $r->formatId;
             
         $tournament->save();
 
         return (new TournamentResource($tournament))
             ->response()
-            ->setStatusCode(202);
+            ->setStatusCode(200);
     }
 
     /**
@@ -138,7 +113,11 @@ class TournamentController extends Controller
      */
     public function show($id)
     {
-        //
+        //Validate $id
+        $v = Validator::make(['id' => $id], ['id' => 'required|integer|exists:tournaments']);
+        if ($v->fails())
+            return $this->errors($v);
+            
         $tournament = Tournament::findOrFail($id);
 
         return new TournamentResource($tournament);
@@ -152,10 +131,14 @@ class TournamentController extends Controller
      */
     public function destroy($id)
     {
-        //
+        //Validate $id
+        $v = Validator::make(['id' => $id], ['id' => 'required|integer|exists:tournaments']);
+        if ($v->fails())
+            return $this->errors($v);
+
         $tournament = Tournament::findOrFail($id)->delete();
 
-        return new TournamentResource($tournament);
+        return response()->setStatusCode(200);
     }
 
 }
